request system complete, next is history button to work in all systems

+++ src/main/java/toshu/org/corpsuite/request/model/Request.java
@@ -38,6 +38,6 @@ public class Request {
 
     private LocalDate toDate;
 
-    private long totalDaysOff;
+    private int totalDaysOff;
 
 }
